refactor Timetable
