+++ src/main/java/com/solvd/schoolschedule/model/Timetable.java
@@ -1,5 +1,7 @@
 package com.solvd.schoolschedule.model;
 
+import com.solvd.schoolschedule.model.interfaces.ITimetableFilter;
+
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
@@ -52,25 +54,6 @@ public class Timetable {
         return object.filter(this);
     }
 
-    public List<Lesson> filterByDay(List<Lesson> lessons, int day) {
-        return lessons.stream()
-                .filter(lesson -> lesson.getTimeSlot().getDay() == day)
-                .sorted(Comparator.comparingInt(l -> l.getTimeSlot().getPeriod()))
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * Get all lessons for a specific teacher on a specific day, sorted by period
-     *
-     * @param teacher the teacher to filter by
-     * @param day     the day to filter by (0-4)
-     * @return list of lessons for the teacher on that day, sorted by period
-     */
-    public List<Lesson> getLessonsForTeacherOnDay(Teacher teacher, int day) {
-        List<Lesson> lessons = this.getLessonsFor(teacher);
-        return filterByDay(lessons, day);
-    }
-
     /**
      * Get all lessons for a specific ITimetableFilter object on a specific day, sorted by period
      *
@@ -78,9 +61,23 @@ public class Timetable {
      * @param day    the day to filter by (0-4)
      * @return list of lessons for the teacher on that day, sorted by period
      */
-    public List<Lesson> getLessonsOnDay(ITimetableFilter object, int day) {
+    public List<Lesson> getLessonsOnDayFor(ITimetableFilter object, int day) {
         List<Lesson> lessons = this.getLessonsFor(object);
         return filterByDay(lessons, day);
+    }
+
+    /**
+     * Filters a lesson list  by a specific day, sorted by period
+     *
+     * @param lessons list of the lessons
+     * @param day     the day to filter by (0-4)
+     * @return list of lessons on that day, sorted by period
+     */
+    private List<Lesson> filterByDay(List<Lesson> lessons, int day) {
+        return lessons.stream()
+                .filter(lesson -> lesson.getTimeSlot().getDay() == day)
+                .sorted(Comparator.comparingInt(l -> l.getTimeSlot().getPeriod()))
+                .collect(Collectors.toList());
     }
 
 }
