+++ src/main/java/com/solvd/schoolschedule/model/interfaces/ITimetableFilter.java
@@ -1,4 +1,7 @@
-package com.solvd.schoolschedule.model;
+package com.solvd.schoolschedule.model.interfaces;
+
+import com.solvd.schoolschedule.model.Lesson;
+import com.solvd.schoolschedule.model.Timetable;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -8,7 +11,7 @@ public interface ITimetableFilter {
     /**
      * Get the ITimetableFilter object from a lesson.
      *
-     * @param lesson
+     * @param lesson lesson
      * @return object ITimetableFilter
      */
     ITimetableFilter getFromLesson(Lesson lesson);
